+++ lib/src/service/model/training_data.dart
@@ -3,11 +3,17 @@ class TrainingData {
   final String baseModel;
   final String language;
   final double wordErrorRate;
+  final double baselineModelWER;
+  final int utterancesInTrain;
+  final int utterancesInDev;
 
   const TrainingData({
     required this.trainingExamples,
     required this.baseModel,
     required this.language,
     required this.wordErrorRate,
+    required this.baselineModelWER,
+    required this.utterancesInDev,
+    required this.utterancesInTrain
   });
 }